LayoutTests/http/tests/contentextensions/main-resource-redirect-error-expected.txt

main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
main frame - didStartProvisionalLoadForFrame
CONSOLE MESSAGE: Content blocker prevented frame displaying http://127.0.0.1:8000/resources/should-not-load.html from loading a resource from http://127.0.0.1:8000/resources/should-not-load.html
main frame - didFailProvisionalLoadWithError
main frame - (kWKErrorCodeFrameLoadBlockedByContentBlocker)

http://127.0.0.1:8000/contentextensions/main-resource-redirect-error.html - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/contentextensions/main-resource-redirect-error.html, main document URL http://127.0.0.1:8000/contentextensions/main-resource-redirect-error.html, http method GET> redirectResponse (null)
http://127.0.0.1:8000/contentextensions/main-resource-redirect-error.html - didReceiveResponse <NSURLResponse http://127.0.0.1:8000/contentextensions/main-resource-redirect-error.html, http status code 200>
http://127.0.0.1:8000/contentextensions/main-resource-redirect-error.html - didFinishLoading
http://127.0.0.1:8000/resources/redirect.py?url=should-not-load.html - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/resources/redirect.py?url=should-not-load.html, main document URL http://127.0.0.1:8000/resources/redirect.py?url=should-not-load.html, http method GET> redirectResponse (null)
